add all feature

+++ models/dto/membership_package_dto.go
@@ -7,18 +7,18 @@ import (
 )
 
 type MembershipPackageResponse struct {
-	ID                  uint64 	`json:"id"`
-	AdminID				uint64	`json:"admin_id"`
-	Name                string	`json:"name"`
-	Duration            uint16	`json:"duration"`
-	CreatedAt time.Time	`json:"created_at"`
-	UpdatedAt time.Time	`json:"updated_at"`
-	DeletedAt gorm.DeletedAt	`json:"deleted_at"`
+	ID        uint64         `json:"id"`
+	AdminID   uint64         `json:"admin_id"`
+	Name      string         `json:"name"`
+	Duration  uint32         `json:"duration"`
+	CreatedAt time.Time      `json:"created_at"`
+	UpdatedAt time.Time      `json:"updated_at"`
+	DeletedAt gorm.DeletedAt `json:"deleted_at"`
 }
 
 type MembershipPackageCreate struct {
-	ID                  uint64 	`json:"id"`
-	AdminID				uint64	`json:"admin_id"`
-	Name                string	`json:"name"`
-	Duration            uint16	`json:"duration"`
+	ID       uint64 `json:"id"`
+	AdminID  uint64 `json:"admin_id"`
+	Name     string `json:"name"`
+	Duration uint16 `json:"duration"`
 }
